[ASTERIXDB-3324][COMP][RT] Stabilize columnar filters and projector
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
This patch includes multiple fixes described in ASTERIXDB-3324

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17993
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Contrib: Michael Blow <mblow@apache.org>

+++ asterixdb/asterix-column/src/main/java/org/apache/asterix/column/values/reader/value/key/FloatKeyValueReader.java
@@ -34,7 +34,7 @@ public final class FloatKeyValueReader extends AbstractFixedLengthColumnKeyValue
     }
 
     @Override
-    public double getDouble() {
+    public float getFloat() {
         return FloatPointable.getFloat(value.getByteArray(), value.getStartOffset());
     }
